rename docs package to apidoc
